Better handling of images
Resize images based on some simple parameters relating to screen size. Seems to do a decent job at scaling, although we're not exactly "tablet optimized" here (we'll add side-by-side fragements at some point). Also remove dependency on ScaleImageView for simplicity - this was doing some of the same work as the ImageLoader's scaling during decoding but messing with the ability to use native layout parameters in XML. Probably won't be doing the StaggeredGrid this time around :(