Fix/finish hike search
Was never really finished. Lines up spinner lables with an equal-length array of values; good enough for now. Also tweaks the wording on the search string.

Also fixed some issues in the AttrDatabase (namely, added a "Level" option to the admin so we can actually select that) plus changed a "ft" value to miles.