Switch to street map type by default
Also, display a toast when the offline map is enabled and user is zoomed too far in or out to see the map tiles. It would be better to have some sort of a floating, static button - perhaps which, when clicked, returned the user to the map area and an acceptable zoom - but we'll leave that for a future feature.